requirements.txt
python-dotenv
pymongo
discord
python-a2s
dblpy
requests